+++ scripts/dysflow-access-runner.ps1
@@ -34,6 +34,10 @@ $BackendPassword = $env:DYSFLOW_BACKEND_PASSWORD
 $startupInfo = $null
 
 function New-DaoDbEngine {
+  if ($env:DYSFLOW_MOCK_COM -eq '1') {
+    return Get-MockDaoDbEngine
+  }
+
   $engineCandidates = @(
     "DAO.DBEngine.160",
     "DAO.DBEngine.150",